MAIN MENU + BADASS DEVICE SELECT
-Basic main menu finished
-Device select finished

Too tired to implement a pause menu for reinputting devices during game. I've done my task for the day. Also there is one new bug where Siepie can't see anything on her side of the screen but y'know I didn't make the camera so that's not my problem. Btw this is probably the shitties implementation of all time BUT IT'S DONE.